[1UP]: headers, inline, paragraphs and lists
The writer is starting to take shape. The most important addition today
is the representation of a list. Lists can be very troublesome to input
by hand. It is even more difficult to read it from raw html; even if it
is formatted properly. The file `test_list.py` shows an example of an
html list and its lexor part.