[NEW]: lexor list
A lexor list as parsed by the default lexor language creates
`list_item` nodes. The addition of the `LexorListItemNW` object
handles displaying the list as it was parsed. Since lists modify the
indentation and Headers are block elements we must set the indentation
level to zero while we write headers. The paragraphs play an important
role with lists, the parser may add the property `remove`, if this
property is set to true this means that a converter must remove the `p`
element and leave its children in place. For the writer it just means
that we will treat it as it if was not there.